Play Page

+++ api/game.py
@@ -1,10 +1,10 @@
 from werkzeug.exceptions import BadRequest, Conflict, NotFound
 
 from flask_restful.reqparse import Argument
+from flask import Blueprint, jsonify
 from flask_socketio import emit
-from flask import Blueprint
 
-from dao import OutgoingGames, Game, User, AuthMethods, RatingArchive
+from dao import GameSettings, OutgoingGames, Game, User, AuthMethods, RatingArchive
 from util import requires_args, requires_auth
 from extensions import CONFIG
 from app import db
@@ -12,7 +12,7 @@ from app import db
 game = Blueprint("game", __name__, url_prefix="/game")
 
 @game.route("/pool/start", methods=["POST"])
-@requires_args(Argument("mode", type=str, required=True), Argument("time_control", type=int, required=True))
+@requires_args(Argument("mode", type=str, required=True), Argument("time_control", type=int, required=True), Argument("increment", type=int, required=True))
 @requires_auth()
 def start_pool_game(user : User, args):
     """
@@ -22,23 +22,25 @@ def start_pool_game(user : User, args):
     if user.outgoing_game: raise Conflict("You already have an outgoing game")
 
     # Check if the mode and the time control are valid
-    mode : str = args.mode
+    mode : str = args.mode.lower()
     time_control : int = args.time_control
+    increment : int = args.increment
     if not mode in CONFIG["MODES"]: raise BadRequest("Invalid Mode")
     if not time_control in CONFIG["TIME_CONTROLS"]: raise BadRequest("Invalid Time Control")
+    if not increment in CONFIG["INCREMENTS"]: raise BadRequest("Invalid Increment")
 
     rating : RatingArchive = user.rating(mode)
 
     # Loop over the game pool, if a game with a similar rating is in the pool, start the game
-    pool : list[OutgoingGames] = OutgoingGames.query.filter_by(mode=mode, time_control=time_control, is_pool=True).order_by(OutgoingGames.created_at).all()
+    pool : list[OutgoingGames] = OutgoingGames.query.filter_by(is_pool=True).order_by(OutgoingGames.created_at).join(GameSettings).filter_by(mode=mode, time_control=time_control, increment=increment).all()
     for outgoing in pool:
-        opponent : User = outgoing.user
+        opponent : User = outgoing.inviter
         if not opponent or \
                 (opponent.auth_method == AuthMethods.GUEST and not user.auth_method == AuthMethods.GUEST): continue
         
         if abs(opponent.rating(mode).elo - rating.elo) <= CONFIG["ACCEPTABLE_RATING_DIFFERENCE"]:
             # Start the game
-            game_id = Game.start_game(user, opponent, mode=mode, time_control=time_control)
+            game_id = Game.start_game(user, opponent, settings=outgoing.game_settings)
             db.session.delete(outgoing)
             db.session.commit()
 
@@ -47,12 +49,13 @@ def start_pool_game(user : User, args):
             return game_id
     
     # If a valid game was not found, add a new game to the pool
-    db.session.add(OutgoingGames(inviter=user, mode=mode, time_control=time_control, is_pool=True))
+    game_settings = GameSettings(mode=mode, time_control=time_control, increment=increment)
+    db.session.add_all([game_settings, OutgoingGames(inviter=user, game_settings=game_settings, is_pool=True)])
     db.session.commit()
-    return "Added to the pool", 200
+    return "Added to the pool", 201
 
 @game.route("/invite", methods=["POST"])
-@requires_args(Argument("recipient_id", type=str, required=True), Argument("mode", type=str, required=True), Argument("time_control", type=int, required=True))
+@requires_args(Argument("recipient_id", type=str, required=True), Argument("mode", type=str, required=True), Argument("time_control", type=int, required=True), Argument("increment", type=int, required=True))
 @requires_auth()
 def invite(user : User, args):
     if user.outgoing_game: raise Conflict("You already have an outgoing game")
@@ -64,13 +67,16 @@ def invite(user : User, args):
     # Check if the mode and the time control are valid
     mode : str = args.mode
     time_control : int = args.time_control
+    increment : int = args.increment
     if not mode in CONFIG["MODES"]: raise BadRequest("Invalid Mode")
     if not time_control in CONFIG["TIME_CONTROLS"]: raise BadRequest("Invalid Time Control")
+    if not increment in CONFIG["INCREMENTS"]: raise BadRequest("Invalid Increment")
 
-    db.session.add(OutgoingGames(inviter=user, recipient=opponent, mode=mode, time_control=time_control))
+    game_settings = GameSettings(mode=mode, time_control=time_control, increment=increment)
+    db.session.add_all([game_settings, OutgoingGames(inviter=user, recipient=opponent, game_settings=game_settings)])
     db.session.commit()
 
-    emit("incoming_game", {"inviter_id": user.user_id}, room=opponent.sid, namespace="/")
+    emit("incoming_games", [user.user_id], room=opponent.sid, namespace="/")
     return "Invite Sent", 200
 
 @game.route("/accept", methods=["POST"])
@@ -80,7 +86,7 @@ def accept_invite(user : User, args):
     inviter : User = User.query.filter_by(user_id=args.inviter_id).first()
     if not inviter or inviter.outgoing_game.recipient != user: raise BadRequest("User Not Invited")
     
-    game_id = Game.start_game(user, inviter, mode=inviter.outgoing_game.mode, time_control=inviter.outgoing_game.time_control)
+    game_id = Game.start_game(user, inviter, settings=inviter.outgoing_game.game_settings)
     db.session.delete(inviter.outgoing_game)
     db.session.commit()
 
@@ -93,5 +99,10 @@ def cancel_outgoing(user : User):
     if not user.outgoing_game: raise NotFound("No Outgoing Games")
 
     db.session.delete(user.outgoing_game)
+    db.session.delete(user.outgoing_game.game_settings)
     db.session.commit()
     return "Request Deleted", 200
+
+@game.route("/has_outgoing", methods=["POST"])
+@requires_auth()
+def has_outgoing(user : User): return jsonify(user.outgoing_game != None)
